feat: profile tasks
