+++ src/components/Profile/UserTasks/UserTasks.module.css
@@ -1,0 +1,14 @@
+.list::-webkit-scrollbar {
+  width: 5px;
+  height: 8px;
+}
+
+.list::-webkit-scrollbar-track {
+  background-color: var(--joy-palette-background-body);
+  border-radius: 5px;
+}
+
+.list::-webkit-scrollbar-thumb {
+  background-color: var(--joy-palette-grey-200);
+  border-radius: 5px;
+}
